Moved back to .module.scss
