+++ packages/styleguide/package.json
@@ -21,17 +21,17 @@
     "@babel/register": "7.8.3",
     "@codecademy/gamut-icons": "^2.1.5",
     "@codecademy/webpack-config": "^4.0.1",
-    "@storybook/addon-a11y": "5.3.9",
-    "@storybook/addon-actions": "5.3.9",
-    "@storybook/addon-centered": "5.3.9",
-    "@storybook/addon-essentials": "5.3.9",
-    "@storybook/addon-info": "5.3.9",
-    "@storybook/addon-knobs": "5.3.9",
-    "@storybook/addon-links": "5.3.9",
-    "@storybook/addon-storyshots": "5.3.9",
-    "@storybook/addons": "5.3.9",
-    "@storybook/react": "5.3.9",
-    "@storybook/theming": "5.3.9",
+    "@storybook/addon-a11y": "5.3.18",
+    "@storybook/addon-actions": "5.3.18",
+    "@storybook/addon-centered": "5.3.18",
+    "@storybook/addon-essentials": "5.3.18",
+    "@storybook/addon-info": "5.3.18",
+    "@storybook/addon-knobs": "5.3.18",
+    "@storybook/addon-links": "5.3.18",
+    "@storybook/addon-storyshots": "5.3.18",
+    "@storybook/addons": "5.3.18",
+    "@storybook/react": "5.3.18",
+    "@storybook/theming": "5.3.18",
     "@types/react": "*",
     "@types/react-color": "^3.0.1",
     "@types/storybook__react": "^5.2.1",
@@ -55,9 +55,5 @@
     "webpack": "4.41.5",
     "webpack-react-docgen-typescript": "0.9.5"
   },
-  "license": "MIT",
-  "devDependencies": {
-    "@types/storybook__addon-knobs": "^5.2.1",
-    "@types/storybook__addon-links": "^5.2.1"
-  }
+  "license": "MIT"
 }
