+++ packages/styleguide/stories/About.stories.mdx
@@ -1,7 +1,7 @@
 import {  Column, LayoutGrid} from '@codecademy/gamut/src';
 import LinkTo from '@storybook/addon-links/react';
 
-import styles from './styles.scss';
+import styles from './styles.module.scss';
 import { AboutCell, AboutHeading } from './Abouts';
 
 <Meta title="About" />
