+++ packages/styleguide/stories/Atoms/FormInputs/FormInputs.stories.tsx
@@ -15,7 +15,7 @@ import {
   StoryTemplate,
   StoryDescription,
 } from '../../Templating';
-import styles from './styles.scss';
+import styles from './styles.module.scss';
 
 export default decoratedStory('Atoms', 'Form Inputs');
