+++ packages/styleguide/stories/Atoms/Icons/styles.module.scss
