+++ packages/styleguide/stories/Brand/Avatar/Avatar.stories.tsx
@@ -8,7 +8,7 @@ import {
   StoryStatus,
   StoryTemplate,
 } from '../../Templating';
-import styles from './styles.scss';
+import styles from './styles.module.scss';
 
 export default decoratedStory('Brand', Avatar);
