+++ packages/styleguide/stories/Brand/Avatar/styles.module.scss
