+++ packages/styleguide/stories/Molecules/Banner/Banner.stories.tsx
@@ -8,7 +8,7 @@ import {
   StoryStatus,
   StoryTemplate,
 } from '../../Templating';
-import styles from './styles.scss';
+import styles from './styles.module.scss';
 
 export default decoratedStory('Molecules', Banner);
