+++ packages/styleguide/stories/Molecules/SplitInterstitial/styles.module.scss
